paas-ce/websocket/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2019.11.28
chardet==3.0.4
Django==2.2.6
django-celery==3.2.1
httplib2==0.9.1
idna==2.8
kombu==3.0.37
Mako==1.0.6
MarkupSafe==1.1
mysqlclient==1.4.4
python-json-logger==0.1.7
pytz==2019.3
six==1.11.0
sqlparse==0.3.0
requests==2.25.0
urllib3==1.26.2
whitenoise==3.3.0
django-cors-headers==3.2.1
django-redis==4.11.0
#terminal
bcrypt==3.2.0
cryptography==3.3.1
paramiko==2.7.2
channels==2.1.4
channels-redis==2.3.1
daphne==2.2.1
#elfinder
Pillow==4.3.0
PyYAML==5.3.1
python-magic==0.4.3
django-extensions==3.1.1
xlrd==1.2.0
pycryptodomex==3.9.8
